Fix burns
svn:sync-xref-src-repo-rev: r138043

+++ src/VBox/Devices/VirtIO/Virtio_1_0.cpp
@@ -1,4 +1,4 @@
-/* $Id: Virtio_1_0.cpp 84353 2020-05-19 05:37:05Z noreply@oracle.com $ */
+/* $Id: Virtio_1_0.cpp 84354 2020-05-19 06:16:04Z noreply@oracle.com $ */
 /** @file
  * Virtio_1_0 - Virtio Common (PCI, feature & config mgt, queue mgt & proxy, notification mgt)
  */
@@ -2143,7 +2143,7 @@ int virtioCoreR3Init(PPDMDEVINS pDevIns, PVIRTIOCORE pVirtio, PVIRTIOCORECC pVir
      * 'unknown' device-specific capability without querying the capability to figure
      *  out size, so pad with an extra page
      */
-    int cbSize = RTStrPrintf(pVirtioCC->pcszMmioName, sizeof(pVirtioCC->pcszMmioName), "%s MMIO", pcszInstance);
+    size_t cbSize = RTStrPrintf(pVirtioCC->pcszMmioName, sizeof(pVirtioCC->pcszMmioName), "%s MMIO", pcszInstance);
     if (cbSize <= 0)
         return PDMDEV_SET_ERROR(pDevIns, rc, N_("virtio: out of memory allocating string")); /* can we put params in this error? */
